stt/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# MinIO client
minio==7.2.0

# Speech recognition
openai-whisper==20231117
webrtcvad==2.0.10
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3
torch==2.1.1
torchaudio==2.1.1

# Audio processing
pydub==0.25.1
scipy==1.11.4

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code quality
ruff==0.1.6
black==23.11.0
isort==5.12.0
mypy==1.7.1
